Game now advances to the next player after a play was successful
Players can now skip turns and cards are removed from a players hand if it is successfully played.

Logic moved from View to Model for handling player out of cards.  Reordered Alleys class (view) so cards are displayed first and then marbles are selected

Created Dealer class that is responsible for dealing cards to players

Moved Deck and shuffling logic out of AlleysGame and into the Dealer class