cleaned up messages some; ensured we had a message in the properties file for each string index, and eliminated some messages that were unused.
set some methods to private that didn't need to be public.

renamed a couple of methods to make their purpose clearer.

implemented checking for end-of-game and start-to-move-partner-
marbles in the Board class

corrected queen's move value to 12 instead of 11

Moved the list of marble spot indices from board (indexed by
color) to Player, to facilitate end-of-game logic.

implemented getPlayingColor() for player; returns the player's
color until all his marbles are in homebase, then his partner's
color if he has a partner. When a player finishes his marbles,
he can start to move his partner's marbles (in the current fake list
of players).

Corrected a bug that allowed a user to move a marble not his own.

Replaced partner color reference within Player with a Player reference
for the partner. Facilitates end-of-game logic.