Mostly documentation additions, some small code changes.
Some of these files are marked by eclipse/git as different, but seem to
have no differences. I'm committing them to get the actual changes on
the server, and to clear whatever flag indicates they need committing. I
think a major reorganization of classes is coming with the client/server
version of Alleys anyway, so I expect either a different repository or
major changes to this one will result.